requirements.txt
python-docx==0.8.11
requests==2.31.0
python-dotenv==1.0.0
pyautogui==0.9.54
keyboard==0.13.5
psutil==7.0.0
openai>=1.0.0
pyperclip>=1.8.2
pywin32>=306; platform_system=="Windows"
Pillow>=10.0.0  # Python Imaging Library for cursor icon
PyAutoIt>=0.6.0; platform_system=="Windows"  # Optional: Python bindings for AutoIt
transformers>=4.24.0  # Hugging Face Transformers for AI model integration
nltk>=3.7  # Natural Language Toolkit for text processing
numpy>=1.23.0  # Numerical Computing Library for data processing
pandas>=1.4.3  # Data Manipulation and Analysis Library
scikit-learn>=1.1.1  # Machine Learning Library for model training
scipy>=1.9.1  # Scientific Computing Library for data analysis
